Call OpenAI with PDF and get affiliations (#6)
Summary:
This program one argument: pdf path

It calls extract_first_page_of_pdf.py with the specified pdf path and a temporary output path. It will save the modified first-page pdf as a temporary file.

Then, it calls get_affiliation with the temporary file which will create a prompt for gpt4.1 to extract each affiliations and return it as a string.

Co-authored-by: Maya V. Nedkova <mayavnedkova@gmail.com>