Get_Coords_From_Affiliation
Includes one function get_coords_from_affiliation which takes an affiliation as a string as its only parameter, calls the geocoord api, and returns a tuple with the latitude and longitude.

Notable mention, this API also allows for the calculation of distance between locations which may be useful in the future.

Fixed typing issue.