Insert affiliation into Postgresql table
Contains two functions, handle_conflict and main.

Main will open the connection to the cloud postgresql, and then for each affiliation in a pre-defined list will call get_coords_from_affiliation and then handle_conflict. handle_conflict will check whether the affiliation already exists, and if so, offer the user the option to update or keep as is. If there is no conflict, it will insert it into the table.

This code is untested, since the tables are not set-up yet for me on the server.