nearest_city into db_to_json (#40)
* Database to JSON

This script will access our data, extract all the needed information for the MAP only, and output it to a json file(for every conference).

I have also included documentation for the json format, but we do not necessarily need to have it.

* nearest_city into db_to_json

Geometric median and find nearest city were updated to work into the db to json script.

Some formatting and other updates applied too.